v0.1.3
removed from version "1" since this is still in development and not a final release.

modified some methods to match std library names.

added some extra methods: reserve, swap, == and !=. note that reserve reserves "enough" for the requested number of data points; that space may not actually be needed to store that information, however (if UTF-8 or UTF-16, less space is likely needed).

added compatibiliy with C++20, allowing char8_t strings to be used. this includes "u8" literals in c++20 whereas those are char before C++20.
note that this relies on the "__cplusplus" macro so this MUST be defined to use C++20. this may need to be "opted-in" with certain compilers (e.g. MSVC).
If you don't use C++20, this should be just fine.